super_benchmark/pyproject.toml
[tool.poetry]
name = "super-benchmark"
version = "0.1.0"
description = ""
authors = ["Acutapugione <acuta.pugione@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.2"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
sqlmodel = "^0.0.22"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
start = "super_benchmark:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
